pyproject.toml
[tool.poetry]
name = "htmxango"
version = "0.1.0"
description = ""
authors = ["Mounir Messelmeni <messelmeni.mounir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.1"
django-htmx = "^1.17.2"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2023.10"
django-render-block = "^0.9.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.0"
black = "^24.1.1"
djlint = "^1.34.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.djlint]
blank_line_after_tag = "load,extends,include"
extend_exclude = "generated_docs"
format_css = false
profile = "django"
use_gitignore = true
ignore = "H025,H021"
